requirements.txt
feast[redis]==0.33.1
kafka-python==2.0.2
mlflow==2.7.1
pandas==1.5.3
scikit-learn==1.3.1
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.13
numpy<2.0.0
dask==2023.10.1
streamlit
plotly
requests